// web/src/app/api/generated/src/models/Preview.ts
export interface Preview {
  /**
   *
   * @type {Array<PublicAnatomyContributionView>}
   * @memberof Preview
   */
  anatomy?: Array<PublicAnatomyContributionView>;
  /**
   *
   * @type {string}
   * @memberof Preview
   */
  environment?: string;
  /**
   *
   * @type {string}
   * @memberof Preview
   */
  exerciseId?: string;
  /**
   *
   * @type {string}
   * @memberof Preview
   */
  exerciseType?: string;
  /**
   *
   * @type {string}
   * @memberof Preview
   */
  exerciseVersionId?: string;
  /**
   *
   * @type {string}
   * @memberof Preview
   */
  instruction?: string;
  /**
   *
   * @type {string}
   * @memberof Preview
   */
  mediaReference?: string;
  /**
   *
   * @type {Array<string>}
   * @memberof Preview
   */
  movementPatterns?: Array<string>;
  /**
   *
   * @type {Array<string>}
   * @memberof Preview
   */
  requiredEquipment?: Array<string>;
  /**
   *
   * @type {string}
   * @memberof Preview
   */
  technicalLevel?: string;
  /**
   *
   * @type {string}
   * @memberof Preview
   */
  title?: string;
  /**
   *
   * @type {number}
   * @memberof Preview
   */
  versionNumber?: number;
}