// web/src/app/api/generated/src/models/Result.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Motion Ecosystem API
 * API for participant and specialist workflows
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Anatomy } from './Anatomy';
import {
  AnatomyFromJSON,
  AnatomyFromJSONTyped,
  AnatomyToJSON,
  AnatomyToJSONTyped,
} from './Anatomy';

/**
 *
 * @export
 * @interface Result
 */
export interface Result {
  /**
   *
   * @type {Array<string>}
   * @memberof Result
   */
  equipment?: Array<string>;
  /**
   *
   * @type {string}
   * @memberof Result
   */
  exerciseId?: string;
  /**
   *
   * @type {string}
   * @memberof Result
   */
  exerciseType?: string;
  /**
   *
   * @type {string}
   * @memberof Result
   */
  exerciseVersionId?: string;
  /**
   *
   * @type {Array<Anatomy>}
   * @memberof Result
   */
  keyAnatomy?: Array<Anatomy>;
  /**
   *
   * @type {string}
   * @memberof Result
   */
  mediaReference?: string;
  /**
   *
   * @type {Array<string>}
   * @memberof Result
   */
  movementPatterns?: Array<string>;
  /**
   *
   * @type {boolean}
   * @memberof Result
   */
  selectable?: boolean;
  /**
   *
   * @type {string}
   * @memberof Result
   */
  summary?: string;
  /**
   *
   * @type {string}
   * @memberof Result
   */
  technicalLevel?: string;
  /**
   *
   * @type {string}
   * @memberof Result
   */
  title?: string;
  /**
   *
   * @type {number}
   * @memberof Result
   */
  versionNumber?: number;
}

/**
 * Check if a given object implements the Result interface.
 */
export function instanceOfResult(value: object): value is Result {
  return true;
}

export function ResultFromJSON(json: any): Result {
  return ResultFromJSONTyped(json, false);
}

export function ResultFromJSONTyped(json: any, ignoreDiscriminator: boolean): Result {
  if (json == null) {
    return json;
  }
  return {
    equipment: json['equipment'] == null ? undefined : json['equipment'],
    exerciseId: json['exerciseId'] == null ? undefined : json['exerciseId'],
    exerciseType: json['exerciseType'] == null ? undefined : json['exerciseType'],
    exerciseVersionId: json['exerciseVersionId'] == null ? undefined : json['exerciseVersionId'],
    keyAnatomy:
      json['keyAnatomy'] == null
        ? undefined
        : (json['keyAnatomy'] as Array<any>).map(AnatomyFromJSON),
    mediaReference: json['mediaReference'] == null ? undefined : json['mediaReference'],
    movementPatterns: json['movementPatterns'] == null ? undefined : json['movementPatterns'],
    selectable: json['selectable'] == null ? undefined : json['selectable'],
    summary: json['summary'] == null ? undefined : json['summary'],
    technicalLevel: json['technicalLevel'] == null ? undefined : json['technicalLevel'],
    title: json['title'] == null ? undefined : json['title'],
    versionNumber: json['versionNumber'] == null ? undefined : json['versionNumber'],
  };
}

export function ResultToJSON(json: any): Result {
  return ResultToJSONTyped(json, false);
}

export function ResultToJSONTyped(
  value?: Result | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    equipment: value['equipment'],
    exerciseId: value['exerciseId'],
    exerciseType: value['exerciseType'],
    exerciseVersionId: value['exerciseVersionId'],
    keyAnatomy:
      value['keyAnatomy'] == null
        ? undefined
        : (value['keyAnatomy'] as Array<any>).map(AnatomyToJSON),
    mediaReference: value['mediaReference'],
    movementPatterns: value['movementPatterns'],
    selectable: value['selectable'],
    summary: value['summary'],
    technicalLevel: value['technicalLevel'],
    title: value['title'],
    versionNumber: value['versionNumber'],
  };
}