// web/src/app/api/generated/src/models/SearchRequest.ts
export interface SearchRequest {
  /**
   *
   * @type {Array<string>}
   * @memberof SearchRequest
   */
  anatomyStructureIds?: Array<string>;
  /**
   *
   * @type {Array<string>}
   * @memberof SearchRequest
   */
  anatomyStructureTypes?: Array<string>;
  /**
   *
   * @type {string}
   * @memberof SearchRequest
   */
  cursor?: string;
  /**
   *
   * @type {Array<string>}
   * @memberof SearchRequest
   */
  equipment?: Array<string>;
  /**
   *
   * @type {number}
   * @memberof SearchRequest
   */
  limit?: number;
  /**
   *
   * @type {string}
   * @memberof SearchRequest
   */
  locale?: string;
  /**
   *
   * @type {Array<string>}
   * @memberof SearchRequest
   */
  movementPatterns?: Array<string>;
  /**
   *
   * @type {Array<string>}
   * @memberof SearchRequest
   */
  positionCodes?: Array<string>;
  /**
   *
   * @type {Array<string>}
   * @memberof SearchRequest
   */
  purposes?: Array<string>;
  /**
   *
   * @type {string}
   * @memberof SearchRequest
   */
  query?: string;
  /**
   *
   * @type {SearchRequestSortEnum}
   * @memberof SearchRequest
   */
  sort?: SearchRequestSortEnum;
  /**
   *
   * @type {Array<string>}
   * @memberof SearchRequest
   */
  technicalLevels?: Array<string>;
  /**
   *
   * @type {boolean}
   * @memberof SearchRequest
   */
  unilateral?: boolean;
}

/**
 * @export
 */
export const SearchRequestSortEnum = {
  Relevance: 'RELEVANCE',
  Name: 'NAME',
  RecentlyPublished: 'RECENTLY_PUBLISHED',
} as const;
export type SearchRequestSortEnum =
  (typeof SearchRequestSortEnum)[keyof typeof SearchRequestSortEnum];